backend/modules/fasta_builder.py
from pathlib import Path
import os

class FastaToPDBBuilder:
    def __init__(self, fasta_file, output_dir="downloads/fasta_built_structures"):
        self.fasta_file = Path(fasta_file)
        self.output_dir = Path(output_dir)
        self.output_dir.mkdir(parents=True, exist_ok=True)

    def build(self):
        from modeller import environ
        from modeller.automodel import automodel

        env = environ()
        env.io.atom_files_directory = [str(self.output_dir)]

        # Define model class
        a = automodel(env,
                      alnfile='alignment.ali',
                      knowns='template',  # dummy template, you can align to real ones
                      sequence='target')
        a.starting_model = 1
        a.ending_model = 1
        a.make()

        print("[INFO] Structure built using Modeller.")

# Usage placeholder:
# builder = FastaToPDBBuilder("downloads/fasta_inputs/sample.fasta")
# builder.build()
