requirements.txt
requests
biopython
openbabel
rdkit
numpy
scipy
matplotlib
pandas